Track Rust master
• unused parameters are now compile errors
• us/is → usize/isize
• convert some usize/isize to u32
• x[] → x[..]
• update feature gates
• thread::Thread::spawn → thread::spawn
• f64::{MIN_VALUE, MAX_VALUE} → f64::{MIN, MAX}
• use nicer range syntax 0u32..16

+++ src/my_scene/bunny.rs
@@ -13,7 +13,7 @@ use scene::{Camera, Scene};
 use vec3::Vec3;
 
 // 300 polys, octree is slightly slower than no octree
-pub fn get_camera(image_width: isize, image_height: isize, fov: f64) -> Camera {
+pub fn get_camera(image_width: u32, image_height: u32, fov: f64) -> Camera {
     Camera::new(
         Vec3 { x: 0.0, y: -150.0, z: 30.0 },
         Vec3 { x: 0.0, y: 60.0, z: 50.0 },
